requirements.txt
neo4j>=5.15.0
langchain-google-genai
pinecone
pyvis==0.3.1
networkx==3.1
tqdm
python-dotenv
langchain
langchain-huggingface 
sentence_transformers
groq